requirements.txt
PyPDF2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.8
tqdm>=4.65.0
Pillow>=8.4.0